refactor(label): migrate labelAttribute and related functionality to sdg-common
