+++ packages/carto/ol/src/full-map/browser/full-map-browser.component.ts
@@ -22,7 +22,7 @@ import {
   ZoomButtonComponent,
   resolveOptions
 } from '@igo2/sdg-carto';
-import { labelAttribute } from '@igo2/sdg-i18n';
+import { labelAttribute } from '@igo2/sdg-common';
 
 import Layer from 'ol/layer/Layer';
