+++ packages/carto/ol/src/reference-map/reference-map-interactions.directive.ts
@@ -14,7 +14,7 @@ import {
 } from '@angular/core';
 
 import { IUseTimeout, detectTouchscreen, useTimeout } from '@igo2/sdg-carto';
-import { labelAttribute } from '@igo2/sdg-i18n';
+import { labelAttribute } from '@igo2/sdg-common';
 
 import MapBrowserEvent from 'ol/MapBrowserEvent';
 import { platformModifierKeyOnly } from 'ol/events/condition';
