+++ packages/carto/package.json
@@ -40,8 +40,7 @@
     "@angular/material": "^21.0.0",
     "@angular/platform-browser": "^21.0.0",
     "@igo2/sdg-core": "*",
-    "@igo2/sdg-i18n": "*",
-    "@igo2/geo": "^21.0.0-next.1",
+    "@igo2/sdg-common": "*",
     "ol": "^10.0.0",
     "rxjs": "^7.0.0"
   }
