+++ packages/carto/src/shared/basemap-switcher/basemap-switcher.component.ts
@@ -8,7 +8,7 @@ import { MatButtonModule } from '@angular/material/button';
 import { MatIconModule } from '@angular/material/icon';
 import { MatTooltipModule } from '@angular/material/tooltip';
 
-import { labelAttribute } from '@igo2/sdg-i18n';
+import { labelAttribute } from '@igo2/sdg-common';
 
 export interface BasemapSwitcherLabels {
   tooltip: string;
