+++ packages/carto/src/shared/navigation/home/home-button.component.ts
@@ -3,7 +3,7 @@ import { MatButtonModule } from '@angular/material/button';
 import { MatIconModule } from '@angular/material/icon';
 import { MatTooltipModule } from '@angular/material/tooltip';
 
-import { labelAttribute } from '@igo2/sdg-i18n';
+import { labelAttribute } from '@igo2/sdg-common';
 
 import { ISdgMap } from '../../map/map.interface';
 import { HomeLabels, IHomeOptions } from './home-button.interface';
