+++ packages/common/src/shared/index.ts
@@ -1,2 +1,3 @@
-export * from './with-labels';
+export * from './label/with-labels';
+export * from './label';
 export * from './anchor/anchor.component';
