+++ packages/common/src/shared/label/index.ts
@@ -1,0 +1,2 @@
+export * from './label';
+export * from './with-labels';
