+++ packages/i18n/src/public-api.ts
@@ -7,6 +7,5 @@ export * from './translation.interface';
 export * from './translation.provider';
 export * from './translation.service';
 export * from './translation.utils';
-export * from './utils';
 
 export { TranslatePipe as TranslationPipe } from '@ngx-translate/core';
