--- packages/i18n/src/utils/index.ts
@@ -1,1 +1,0 @@
-export * from './coercion';
